refactor: restructure project
