+++ beine_cpp/include/beine_cpp/provider/stance_provider.hpp
@@ -18,8 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef BEINE_CPP__LEGS__JOINTS_CONSUMER_HPP_
-#define BEINE_CPP__LEGS__JOINTS_CONSUMER_HPP_
+#ifndef BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
+#define BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
 
 #include <rclcpp/rclcpp.hpp>
 
@@ -28,66 +28,60 @@
 namespace beine_cpp
 {
 
-class JointsConsumer
+class StanceProvider
 {
 public:
-  inline JointsConsumer();
-  inline explicit JointsConsumer(rclcpp::Node::SharedPtr node);
+  inline StanceProvider();
+  inline explicit StanceProvider(rclcpp::Node::SharedPtr node);
 
   inline void set_node(rclcpp::Node::SharedPtr node);
 
-  inline rclcpp::Node::SharedPtr get_node() const;
+  inline void set_stance(Stance stance);
 
-  inline const Joints & get_joints() const;
+  inline rclcpp::Node::SharedPtr get_node();
 
 private:
   rclcpp::Node::SharedPtr node;
 
-  rclcpp::Subscription<Joints>::SharedPtr joints_subscription;
-
-  Joints current_joints;
+  rclcpp::Publisher<StanceMsg>::SharedPtr stance_publisher;
 };
 
-JointsConsumer::JointsConsumer()
+StanceProvider::StanceProvider()
 {
 }
 
-JointsConsumer::JointsConsumer(rclcpp::Node::SharedPtr node)
-: JointsConsumer()
+StanceProvider::StanceProvider(rclcpp::Node::SharedPtr node)
+: StanceProvider()
 {
   set_node(node);
 }
 
-void JointsConsumer::set_node(rclcpp::Node::SharedPtr node)
+void StanceProvider::set_node(rclcpp::Node::SharedPtr node)
 {
   // Initialize the node
   this->node = node;
 
-  // Initialize the joints subscription
+  // Initialize the stance publisher
   {
-    joints_subscription = get_node()->create_subscription<Joints>(
-      "/legs/joints", 10,
-      [this](const Joints::SharedPtr joints) {
-        current_joints = *joints;
-      });
+    stance_publisher = get_node()->create_publisher<StanceMsg>("/legs/stance", 10);
 
     RCLCPP_INFO_STREAM(
       get_node()->get_logger(),
-      "Joints subscription initialized on " <<
-        joints_subscription->get_topic_name() << "!");
+      "Stance publisher initialized on " <<
+        stance_publisher->get_topic_name() << "!");
   }
 }
 
-rclcpp::Node::SharedPtr JointsConsumer::get_node() const
+void StanceProvider::set_stance(Stance stance)
+{
+  stance_publisher->publish(stance);
+}
+
+rclcpp::Node::SharedPtr StanceProvider::get_node()
 {
   return node;
 }
 
-const Joints & JointsConsumer::get_joints() const
-{
-  return current_joints;
-}
-
 }  // namespace beine_cpp
 
-#endif  // BEINE_CPP__LEGS__JOINTS_CONSUMER_HPP_
+#endif  // BEINE_CPP__PROVIDER__STANCE_PROVIDER_HPP_
